Add different scraper services for different websites
