+++ app/services/remote_ok_scraper.rb
@@ -1,0 +1,5 @@
+# app/services/remote_ok_scraper.rb
+class RemoteOkScraper < BaseScraper
+  def parse(document)
+  end
+end
